`linera-service`: compile with `linera-client`
