+++ linera-client/src/chain_listener.rs
@@ -89,7 +89,7 @@ where
     ViewError: From<S::StoreError>,
 {
     /// Creates a new chain listener given client chains.
-    pub(crate) fn new(config: ChainListenerConfig, clients: ChainClients<P, S>) -> Self {
+    pub fn new(config: ChainListenerConfig, clients: ChainClients<P, S>) -> Self {
         Self { config, clients }
     }
