+++ linera-client/src/client_options.rs
@@ -129,17 +129,16 @@ impl ClientOptions {
         }
     }
 
-    pub async fn run_command_with_storage<R: Runnable>(self, job: R) -> anyhow::Result<R::Output> {
+    pub async fn run_with_storage<R: Runnable>(&self, job: R) -> anyhow::Result<R::Output> {
         let wallet = self.wallet()?;
-        let output = run_with_storage(
+        let output = Box::pin(run_with_storage(
             self.storage_config()?
                 .add_common_config(self.common_config())
                 .await?,
-            &wallet.genesis_config().clone(),
+            &genesis_config,
             self.wasm_runtime.with_wasm_default(),
             job,
-        )
-        .boxed()
+        ))
         .await?;
         Ok(output)
     }
