+++ linera-client/src/lib.rs
@@ -3,6 +3,7 @@ pub mod chain_listener;
 pub mod client_context;
 pub mod client_options;
 pub mod config;
+pub mod persistent;
 pub mod storage;
 pub mod util;
 pub mod wallet;
