+++ linera-service/Cargo.toml
@@ -15,21 +15,28 @@ version.workspace = true
 ethereum = []
 default = ["wasmer", "rocksdb"]
 test = ["linera-views/test", "linera-execution/test", "dep:stdext"]
-benchmark = ["linera-base/test", "dep:linera-sdk"]
-wasmer = ["linera-execution/wasmer", "linera-storage/wasmer"]
-wasmtime = ["linera-execution/wasmtime", "linera-storage/wasmtime"]
+benchmark = ["linera-base/test", "linera-client/benchmark", "dep:linera-sdk"]
+wasmer = ["linera-client/wasmer", "linera-execution/wasmer", "linera-storage/wasmer"]
+wasmtime = ["linera-client/wasmtime", "linera-execution/wasmtime", "linera-storage/wasmtime"]
 rocksdb = [
+    "linera-client/rocksdb",
     "linera-views/rocksdb",
     "linera-core/rocksdb",
     "linera-storage/rocksdb",
 ]
-dynamodb = ["linera-views/dynamodb", "linera-core/dynamodb", "linera-storage/dynamodb"]
+dynamodb = [
+    "linera-client/dynamodb",
+    "linera-views/dynamodb",
+    "linera-core/dynamodb",
+    "linera-storage/dynamodb",
+]
 scylladb = [
+    "linera-client/scylladb",
     "linera-views/scylladb",
     "linera-core/scylladb",
     "linera-storage/scylladb",
 ]
-kubernetes = ["dep:k8s-openapi", "dep:kube", "dep:pathdiff", "dep:fs_extra"]
+kubernetes = ["linera-client/kubernetes", "dep:k8s-openapi", "dep:kube", "dep:pathdiff", "dep:fs_extra"]
 remote_net = ["dep:k8s-openapi", "dep:kube"]
 metrics = ["prometheus", "linera-base/metrics"]
 storage_service = []
@@ -62,6 +69,7 @@ k8s-openapi = { workspace = true, optional = true }
 kube = { workspace = true, optional = true }
 linera-base = { workspace = true, features = ["metrics"] }
 linera-chain = { workspace = true, features = ["metrics"] }
+linera-client = { workspace = true, features = ["metrics"] }
 linera-core = { workspace = true, features = ["metrics", "rocksdb", "wasmer"] }
 linera-execution = { workspace = true, features = ["fs", "metrics", "wasmer"] }
 linera-rpc = { workspace = true, features = ["server", "simple-network"] }
@@ -107,6 +115,7 @@ fungible.workspace = true
 linera-alloy = { workspace = true, default-features = false, features = [ "rpc-types-eth" ] }
 linera-base = { workspace = true, features = ["test"] }
 linera-chain = { workspace = true, features = ["test"] }
+linera-client = { workspace = true, features = ["test"] }
 linera-core = { workspace = true, features = ["test"] }
 linera-ethereum.workspace = true
 linera-execution = { workspace = true, features = ["test"] }
