+++ linera-service/src/cli_wrappers/local_net.rs
@@ -33,9 +33,10 @@ use {linera_views::rocks_db::create_rocks_db_test_path, std::ops::Deref};
 
 use crate::{
     cli_wrappers::{ClientWrapper, LineraNet, LineraNetConfig, Network},
-    storage::{StorageConfig, StorageConfigNamespace},
     util::ChildExt,
 };
+
+use linera_client::storage::{StorageConfig, StorageConfigNamespace};
 
 #[cfg(all(feature = "rocksdb", with_testing))]
 trait LocalServerInternal: Sized {
