+++ linera-service/src/faucet.rs
@@ -27,7 +27,9 @@ use thiserror::Error as ThisError;
 use tower_http::cors::CorsLayer;
 use tracing::{error, info};
 
-use crate::{chain_listener::ClientContext, config::GenesisConfig, util};
+use linera_client::{chain_listener::ClientContext, config::GenesisConfig};
+
+use crate::util;
 
 #[cfg(test)]
 #[path = "unit_tests/faucet.rs"]
